### 2023-09-29 #5900 @editorialbot
[PRE REVIEW]: GTFS Segments: A fast and efficient library to generate bus stop spacings
**Submitting author:** @praneethd7 (<a href="http://orcid.org/0000-0002-5911-4681">Saipraneeth Devunuri</a>)
**Repository:** https://github.com/UTEL-UIUC/gtfs_segments
**Branch with paper.md** (empty if default branch):
**Version:** v2.0.1
