+++ paper/paper.bib
@@ -4,7 +4,7 @@
   year          = {2023},
   eprint        = {2208.04394},
   archiveprefix = {arXiv},
-  primaryclass  = {stat.ME}
+  primaryclass  = {stat.ME},
   doi           = {https://doi.org/10.48550/arXiv.2208.04394}
 }
 
@@ -14,7 +14,7 @@
   year          = {2023},
   eprint        = {2308.02618},
   archiveprefix = {arXiv},
-  primaryclass  = {cs.IR}
+  primaryclass  = {cs.IR},
   doi           = {https://doi.org/10.48550/arXiv.2308.02618}
 }
 
@@ -48,14 +48,14 @@
   number    = {3},
   pages     = {453--466},
   year      = {2023},
-  publisher = {Springer}
+  publisher = {Springer},
   doi       = {https://doi.org/10.31235/osf.io/qydr6}
 }
 
 @article{lehe4135394bus,
   title   = {Bus stop spacing with heterogeneous trip lengths and elastic demand},
   author  = {Lehe, Lewis and Pandey, Ayush},
-  journal = {Available at SSRN 4135394}
+  journal = {Available at SSRN 4135394},
   doi     = {https://doi.org/10.2139/ssrn.4135394}
 }
 
@@ -113,6 +113,6 @@
   title   = {Analysis of approximate nearest neighbor searching with clustered point sets},
   author  = {Maneewongvatana, Songrit and Mount, David M},
   journal = {arXiv preprint cs/9901013},
-  year    = {1999}
+  year    = {1999},
   doi     = {https://doi.org/10.48550/arXiv.cs/9901013}
 }
