+++ paper/paper.bib
@@ -5,7 +5,7 @@
   eprint        = {2208.04394},
   archiveprefix = {arXiv},
   primaryclass  = {stat.ME},
-  doi           = {https://doi.org/10.48550/arXiv.2208.04394}
+  doi           = {10.48550/arXiv.2208.04394}
 }
 
 @misc{devunuri2023chatgpt,
@@ -15,7 +15,7 @@
   eprint        = {2308.02618},
   archiveprefix = {arXiv},
   primaryclass  = {cs.IR},
-  doi           = {https://doi.org/10.48550/arXiv.2308.02618}
+  doi           = {10.48550/arXiv.2308.02618}
 }
 
 @data{DVN/SFBIVU_2022,
@@ -26,7 +26,7 @@
   year      = {2022},
   version   = {V3},
   doi       = {10.7910/DVN/SFBIVU},
-  url       = {https://doi.org/10.7910/DVN/SFBIVU}
+  url       = {10.7910/DVN/SFBIVU}
 }
 
 @data{DVN/QFTAPM_2023,
@@ -37,7 +37,7 @@
   year      = {2023},
   version   = {V2},
   doi       = {10.7910/DVN/QFTAPM},
-  url       = {https://doi.org/10.7910/DVN/QFTAPM}
+  url       = {10.7910/DVN/QFTAPM}
 }
 
 @article{pereira2023exploring,
@@ -49,14 +49,14 @@
   pages     = {453--466},
   year      = {2023},
   publisher = {Springer},
-  doi       = {https://doi.org/10.31235/osf.io/qydr6}
+  doi       = {10.31235/osf.io/qydr6}
 }
 
 @article{lehe4135394bus,
   title   = {Bus stop spacing with heterogeneous trip lengths and elastic demand},
   author  = {Lehe, Lewis and Pandey, Ayush},
   journal = {Available at SSRN 4135394},
-  doi     = {https://doi.org/10.2139/ssrn.4135394}
+  doi     = {10.2139/ssrn.4135394}
 }
 
 @software{Toso2023,
@@ -114,5 +114,5 @@
   author  = {Maneewongvatana, Songrit and Mount, David M},
   journal = {arXiv preprint cs/9901013},
   year    = {1999},
-  doi     = {https://doi.org/10.48550/arXiv.cs/9901013}
+  doi     = {10.48550/arXiv.cs/9901013}
 }
