+++ requirements.txt
@@ -14,3 +14,4 @@ utm==0.7.0
 rasterio==1.3.6
 cchardet==2.1.7
 isoweek==1.3.3
+charset_normalizer==3.3.0
