+++ setup.py
@@ -1,6 +1,6 @@
 from setuptools import setup, find_packages
 
-__version__ = '0.0.9'
+__version__ = '0.0.9a'
 DESCRIPTION = 'GTFS segments'
 with open("README.md", "r", encoding="utf8") as fh:
     LONG_DESCRIPTION = fh.read()
@@ -15,7 +15,8 @@ REQUIREMENTS = ['geopandas',
                 'contextily',
                 'requests',
                 'isoweek',
-                'cchardet']
+                'cchardet',
+                'charset_normalizer']
 
 # Setting up
 setup(
