### 2023-10-04 ed962ec
Add typing for remaining files
