+++ gtfs_segments/gtfs_segments.py
@@ -63,9 +63,7 @@ def make_segments_unique(df, traversal_threshold=1) -> gpd.GeoDataFrame:
         for i, index in enumerate(inds):
             if i != 0:
                 seg_split = key[1].split("-")
-                df.loc[index, "segment_id"] = (
-                    seg_split[0] + "-" + seg_split[1] + "-" + str(i + 1)
-                )
+                df.loc[index, "segment_id"] = seg_split[0] + "-" + seg_split[1] + "-" + str(i + 1)
     grp_again = df.groupby(["route_id", "segment_id"])
     df = grp_again.first().reset_index()
     df["traversals"] = grp_again["traversals"].sum().values
@@ -73,7 +71,7 @@ def make_segments_unique(df, traversal_threshold=1) -> gpd.GeoDataFrame:
     return df
 
 
-def filter_stop_df(stop_df, trip_ids) -> gpd.GeoDataFrame:
+def filter_stop_df(stop_df, trip_ids, stop_loc_df) -> gpd.GeoDataFrame:
     """
     It takes a dataframe of stops and a list of trip IDs and returns a dataframe of stops that are in
     the list of trip IDs
@@ -85,6 +83,19 @@ def filter_stop_df(stop_df, trip_ids) -> gpd.GeoDataFrame:
     Returns:
       A dataframe with the trip_id, s top_id, and stop_sequence for the trips in the trip_ids list.
     """
+    missing_stop_locs = set(stop_df.stop_id) - set(stop_loc_df.stop_id)
+    if len(missing_stop_locs) > 0:
+        print("Missing stop locations for:", missing_stop_locs)
+        missing_trips = stop_df[stop_df.stop_id.isin(missing_stop_locs)].trip_id.unique()
+        for trip in missing_trips:
+            trip_ids.discard(trip)
+            print(
+                "Removed the trip_id:",
+                trip,
+                "as stop locations are missing for stops in the trip"
+            )
+    # Filter the stop_df to only include the trip_ids in the trip_ids list
+    stop_df = stop_df[stop_df.trip_id.isin(trip_ids)].reset_index(drop=True)
     stop_df = stop_df.sort_values(["trip_id", "stop_sequence"]).reset_index(drop=True)
     stop_df["main_index"] = stop_df.index
     stop_df_grp = stop_df.groupby("trip_id")
@@ -99,7 +110,7 @@ def filter_stop_df(stop_df, trip_ids) -> gpd.GeoDataFrame:
     if len(drop_inds) > 0 and len(drop_inds[0]) > 0:
         stop_df = stop_df[~stop_df["main_index"].isin(drop_inds)].reset_index(drop=True)
     stop_df = stop_df[["trip_id", "stop_id", "stop_sequence"]]
-    stop_df = stop_df[stop_df.trip_id.isin(trip_ids)].reset_index(drop=True)
+
     stop_df = stop_df.sort_values(["trip_id", "stop_sequence"]).reset_index(drop=True)
     return stop_df
 
@@ -115,9 +126,7 @@ def merge_stop_geom(stop_df, stop_loc_df) -> gpd.GeoDataFrame:
     Returns:
       A GeoDataFrame
     """
-    stop_df["start"] = stop_df.copy().merge(stop_loc_df, how="left", on="stop_id")[
-        "geometry"
-    ]
+    stop_df["start"] = stop_df.copy().merge(stop_loc_df, how="left", on="stop_id")["geometry"]
     stop_df = gpd.GeoDataFrame(stop_df, geometry="start")
     return make_gdf(stop_df)
 
[…]
