### 2023-10-09 8e5fa48
fix Docs| Partridge Arg Typing
