+++ gtfs_segments/partridge_func.py
@@ -1,32 +1,29 @@
 import pandas as pd
 import gtfs_segments.partridge_mod as ptg
 from .partridge_mod.gtfs import Feed
-from typing import Tuple
+from typing import Tuple, All
 
 
-def get_bus_feed(path, agency_id=None, threshold=1) -> Tuple[str, Feed]:
+def get_bus_feed(path: str, agency_id: All = None, threshold: int = 1) -> Tuple[str, Feed]:
     """
-    Retrieve bus feed data from a specified path, with the option to filter by agency name, and
-    return the busiest date and a GTFS feed object.
-
+    The `get_bus_feed` function retrieves bus feed data from a specified path, with the option to filter
+    by agency name, and returns the busiest date and a GTFS feed object.
+    
     Args:
-      path: The path to the GTFS file.
-      agency_id:
-        The `agency_id` parameter is an optional parameter that allows you to filter
-        the bus feed data by the
-      agency name:
-        If you provide an `agency_id`, the function will retrieve the bus feed for
-        that specific transit agency. If you do not provide an `agency_id`, whole feed is used.
-      threshold:
-        The threshold parameter is used to filter out service IDs that have a
-        low frequency. It is set to a default value of 1, but you can adjust it to
-        a different value if needed. Service IDs with a sum of stop times greater
-        than the threshold will be included in the returned bus feed data. Defaults to 1
-
+      path (str): The `path` parameter is a string that represents the path to the GTFS file. This file
+    contains the bus feed data.
+      agency_id (All): The `agency_id` parameter is an optional parameter that allows you to filter the
+    bus feed data by the agency name. It is used to specify the ID of the transit agency for which you
+    want to retrieve the bus feed data. If you provide an `agency_id`, the function will only return
+    data
+      threshold (int): The `threshold` parameter is used to filter out service IDs that have a low
+    frequency. It is set to a default value of 1, but you can adjust it to a different value if needed.
+    Service IDs with a sum of stop times greater than the threshold will be included in the returned
+    bus. Defaults to 1
+    
     Returns:
       A tuple containing the busiest date and a GTFS feed object. The GTFS feed object contains
-      information about routes, stops, stop times, trips, and shapes for a transit agency's
-      schedule.
+    information about routes, stops, stop times, trips, and shapes for a transit agency's schedule.
     """
     _date, bday_service_ids = ptg.read_busiest_date(path)
     all_days_s_ids_df = get_all_days_s_ids(path)
@@ -54,7 +51,7 @@ def get_bus_feed(path, agency_id=None, threshold=1) -> Tuple[str, Feed]:
     return _date, feed
 
 
-def get_all_days_s_ids(path)-> pd.DataFrame:
+def get_all_days_s_ids(path: str) -> pd.DataFrame:
     """
     Read dates by service IDs from a given path, create a DataFrame, populate it with the dates and
     service IDs, and fill missing values with False.
