+++ gtfs_segments/partridge_func.py
@@ -1,10 +1,10 @@
 import pandas as pd
 import gtfs_segments.partridge_mod as ptg
 from .partridge_mod.gtfs import Feed
-from typing import Tuple, All
+from typing import Tuple, Any
 
 
-def get_bus_feed(path: str, agency_id: All = None, threshold: int = 1) -> Tuple[str, Feed]:
+def get_bus_feed(path: str, agency_id: Any = None, threshold: int = 1) -> Tuple[str, Feed]:
     """
     The `get_bus_feed` function retrieves bus feed data from a specified path, with the option to filter
     by agency name, and returns the busiest date and a GTFS feed object.
