+++ gtfs_segments/utils.py
@@ -13,7 +13,9 @@ from typing import Any
 plt.style.use("ggplot")
 
 
-def plot_hist(df, save_fig=False, show_mean=False, **kwargs) -> plt.Figure:
+def plot_hist(
+    df: pd.DataFrame, save_fig: bool = False, show_mean: bool = False, **kwargs
+) -> plt.Figure:
     """
     It takes a dataframe with two columns, one with the distance between stops and the other with the
     number of traversals between those stops, and plots a weighted histogram of the distances
@@ -71,7 +73,9 @@ def plot_hist(df, save_fig=False, show_mean=False, **kwargs) -> plt.Figure:
     return fig
 
 
-def summary_stats(df, max_spacing=3000, export=False, **kwargs) -> pd.DataFrame:
+def summary_stats(
+    df: pd.DataFrame, max_spacing: float = 3000, export: bool = False, **kwargs
+) -> pd.DataFrame:
     """
     It takes in a dataframe, and returns a dataframe with summary statistics
 
@@ -84,25 +88,15 @@ def summary_stats(df, max_spacing=3000, export=False, **kwargs) -> pd.DataFrame:
     """
     print("Using max_spacing = ", max_spacing)
     percent_spacing = round(
-        df[df["distance"] > max_spacing]["traversals"].sum()
-        / df["traversals"].sum()
-        * 100,
+        df[df["distance"] > max_spacing]["traversals"].sum() / df["traversals"].sum() * 100,
         3,
     )
     df = df[df["distance"] <= max_spacing]
     seg_weighted_mean = (
-        df.groupby(["segment_id", "distance"])
-        .first()
-        .reset_index()["distance"]
-        .mean()
-        .round(2)
+        df.groupby(["segment_id", "distance"]).first().reset_index()["distance"].mean().round(2)
     )
     seg_weighted_median = (
-        df.groupby(["segment_id", "distance"])
-        .first()
-        .reset_index()["distance"]
-        .median()
-        .round(2)
+        df.groupby(["segment_id", "distance"]).first().reset_index()["distance"].median().round(2)
     )
     route_weighted_mean = (
         df.groupby(["route_id", "segment_id", "distance"])
@@ -118,9 +112,7 @@ def summary_stats(df, max_spacing=3000, export=False, **kwargs) -> pd.DataFrame:
         .median()
         .round(2)
     )
-    weighted_data = np.hstack(
-        [np.repeat(x, y) for x, y in zip(df["distance"], df.traversals)]
-    )
+    weighted_data = np.hstack([np.repeat(x, y) for x, y in zip(df["distance"], df.traversals)])
 
     df_dict = {
         "Segment Weighted Mean": seg_weighted_mean,
@@ -149,7 +141,9 @@ def summary_stats(df, max_spacing=3000, export=False, **kwargs) -> pd.DataFrame:
     return summary_df
 
 
-def export_segments(df, file_path, output_format, geometry=True) -> None:
+def export_segments(
+    df: pd.DataFrame, file_path: str, output_format: str, geometry: bool = True
+) -> None:
     """
     This function takes a GeoDataFrame of segments, a file path, an output format, and a boolean value
     for whether or not to include the geometry in the output.
@@ -188,7 +182,7 @@ def export_segments(df, file_path, output_format, geometry=True) -> None:
         s_df["start_lat"] = [g[0][1] for g in geom_list]
         s_df["end_lon"] = [g[-1][0] for g in geom_list]
         s_df["end_lat"] = [g[-1][1] for g in geom_list]
-        if geometry == True:
+        if geometry:
             # Output With LS
             sg_df.to_csv(file_path, index=False)
         else:
[…]
