### 2023-10-16 4f5629d
Change default basemap to CartoDB Positron
Stamen deprecated their toner-lite basemap, so we're switching to
CartoDB Positron. This is a more modern basemap, and it's also similar
light color scheme. The default colors are also changed to reflect a
flat UI color scheme. Also  added option to change basemap with
`map_provider` argument. The `dpi` argument is also added to control
the resolution of the output image.

Also added test for route_stats.py
