+++ gtfs_segments/geom_utils.py
@@ -105,42 +105,63 @@ def get_zone_epsg(stop_df) -> int:
 
 
 def view_spacings(
-    df, basemap=False, show_stops=False, level="whole", axis="on", **kwargs
+    df,
+    basemap=False,
+    map_provider=cx.providers.CartoDB.Positron,
+    show_stops=False,
+    level="whole",
+    axis="on",
+    dpi=300,
+    **kwargs
 ) -> plt.Figure:
     """
-    > This function plots the spacings of the bus network, or a specific route,
-    or a specific segment
+    The `view_spacings` function plots the spacings of a bus network, route, or segment, with options to
+    add a basemap and show stops.
 
     Args:
-      df: the dataframe containing the bus network
-      basemap: if True, will add a basemap to the plot. Defaults to False
-      level: "whole" or "route" or "segment". Defaults to whole
-      axis: 'on' or 'off'. Defaults to on
+      df: The GTFS segments dataframe containing the bus network data.
+      basemap: The `basemap` parameter is a boolean value that determines whether to add a basemap to
+    the plot. If set to `True`, a basemap will be added. If set to `False`, no basemap will be added.
+    The default value is `False`. Defaults to False
+      map_provider: The `map_provider` parameter is used to specify the source of the basemap that
+    will be added to the plot. It is set to `cx.providers.CartoDB.Positron` by default, which means
+    that the basemap will be sourced from CartoDB's Positron. Use `contextily.providers` to see full
+    list of providers
+      show_stops: The `show_stops` parameter is a boolean flag that determines whether or not to display
+    the bus stops on the plot. If set to `True`, the bus stops will be shown as white markers on the
+    plot. If set to `False`, the bus stops will not be shown. Defaults to False
+      level: The "level" parameter determines the level of detail to plot the spacings. It can take one
+    of three values:. Defaults to whole
+      axis: The `axis` parameter determines whether the axis of the plot should be displayed or not. If
+    `axis` is set to "on", the axis will be displayed. If `axis` is set to "off", the axis will not be
+    displayed. Defaults to on
+      dpi: The `dpi` parameter determines the resolution of the plot. Defaults to 300
+
+    Returns:
+      a matplotlib Figure object.
     """
-    fig, ax = plt.subplots(figsize=(10, 10), dpi=100)
+    fig, ax = plt.subplots(figsize=(10, 10), dpi=dpi)
     crs = df.crs
     # Filter based on direction and level
     if "direction" in kwargs.keys():
         df = df[df.direction_id == kwargs["direction"]].copy()
     if level == "whole":
-        markersize = 15
+        markersize = 20
         ax = df.plot(
             ax=ax,
-            color="y",
-            linewidth=0.50,
+            color="#34495e",
+            linewidth=0.5,
             edgecolor="black",
             label="Bus network",
             zorder=1,
         )
     elif level == "route":
-        markersize = 30
+        markersize = 40
         assert "route" in kwargs.keys(), "Please provide a route_id in route attibute"
         df = df[df.route_id == kwargs["route"]].copy()
     elif level == "segment":
-        markersize = 50
-        assert (
-            "segment" in kwargs.keys()
-        ), "Please provide a segment_id in segment attibute"
+        markersize = 60
+        assert "segment" in kwargs.keys(), "Please provide a segment_id in segment attibute"
         df = df[df.segment_id == kwargs["segment"]].copy()
     else:
         raise ValueError("level must be either whole, route, or segment")
@@ -151,7 +172,7 @@ def view_spacings(
         ax = df.plot(
             ax=ax,
             linewidth=1.5,
[…]
