+++ README.md
@@ -221,7 +221,7 @@ Each row does not represent one segment. Rather, each row maps to a combination 
 
 ```python
 from gtfs_segments import view_spacings
-view_spacings(segments_df,route = '18131',segment = '6294-6290-1',basemap=True)
+view_spacings(segments_df,route = '8',segment = '6364-3725-1',basemap=True)
 ```
 <div align='center'><a>
     <img src="https://raw.githubusercontent.com/UTEL-UIUC/gtfs_segments/main/images/visualize.jpg" alt="data" width=600>
