+++ test/test_utils.py
@@ -6,10 +6,11 @@ import geopandas as gpd
 import pandas as pd
 import matplotlib
 from gtfs_segments import get_gtfs_segments, plot_hist, summary_stats, export_segments
+
 test_dir = os.path.dirname(__file__)
 
 
-class TestGTFSSegments(unittest.TestCase):
+class TestUtils(unittest.TestCase):
     """Tests for utils.py module in the package."""
 
     def setUp(self):
@@ -34,15 +35,16 @@ class TestGTFSSegments(unittest.TestCase):
             "Error with plot_hist. Should work for the SFMTA example feed",
         )
         fig = plot_hist(
-            self.df, show_mean=True, save_fig=True, file_path="test_hist.png"
+            self.df,
+            show_mean=True,
+            save_fig=True,
+            file_path=os.path.join(test_dir, "output", "test_hist.png"),
         )
         self.assertTrue(
             type(fig) == matplotlib.figure.Figure,
             "Error with plot_hist. Should work for the SFMTA example feed",
         )
-        self.assertTrue(
-            os.path.exists("test_hist.png"), "Check if the test_hist.png file exists"
-        )
+        self.assertTrue(os.path.exists("test_hist.png"), "Check if the test_hist.png file exists")
 
     def test_summary_stats(self):
         summ_df = summary_stats(self.df)
@@ -51,14 +53,16 @@ class TestGTFSSegments(unittest.TestCase):
             "Error with summary_stats. Result should be a DataFrame",
         )
         summ_df = summary_stats(
-            self.df, export=True, file_path="test_summary_stats.csv"
+            self.df,
+            export=True,
+            file_path=os.path.join(test_dir, "output", "test_summary_stats.csv"),
         )
         self.assertTrue(
             isinstance(summ_df, pd.DataFrame),
             "Error with summary_stats. Result should be a DataFrame",
         )
         self.assertTrue(
-            os.path.exists("test_summary_stats.csv"),
+            os.path.exists(os.path.join(test_dir, "output", "test_summary_stats.csv")),
             "Check if the test_summary_stats.csv file exists",
         )
         self.assertTrue(
@@ -68,12 +72,16 @@ class TestGTFSSegments(unittest.TestCase):
 
     def test_export_segments(self):
         # Test export_segments for .csv and with geometry
-        export_segments(self.df, "test_export_segments.csv", output_format="csv")
+        export_segments(
+            self.df,
+            os.path.join(test_dir, "output", "test_export_segments.csv"),
+            output_format="csv",
+        )
         self.assertTrue(
-            os.path.exists("test_export_segments.csv"),
+            os.path.exists(os.path.join(test_dir, "output", "test_export_segments.csv")),
             "Check if the test_export_segments.csv file exists",
         )
-        df = pd.read_csv("test_export_segments.csv")
+        df = pd.read_csv(os.path.join(test_dir, "output", "test_export_segments.csv"))
         self.assertTrue(
             type(df) == pd.DataFrame,
             "Import again to see if the file is valid and has the same data",
@@ -85,30 +93,33 @@ class TestGTFSSegments(unittest.TestCase):
 
         # Test export_segments for .csv and without geometry
         export_segments(
-            self.df, "test_export_segments.csv", output_format="csv", geometry=False
+            self.df,
+            os.path.join(test_dir, "output", "test_export_segments.csv"),
+            output_format="csv",
+            geometry=False,
         )
         self.assertTrue(
-            os.path.exists("test_export_segments.csv"),
+            os.path.exists(os.path.join(test_dir, "output", "test_export_segments.csv")),
             "Check if the test_export_segments.csv file exists",
         )
-        df = pd.read_csv("test_export_segments.csv")
[…]
