### 2023-10-18 2a82200
Create seperate requirements for docs
