+++ .readthedocs.yaml
@@ -16,4 +16,4 @@ mkdocs:
 # Optionally declare the Python requirements required to build your docs
 python:
   install:
-  - requirements: docs/requirements.txt
+  - requirements: docs/requirements_docs.txt
