+++ docs/requirements_docs.txt
@@ -1,22 +1,0 @@
-cython==3.0.2
-contextily==1.2.0
-geopandas==0.12.2
-isoweek==1.3.3
-matplotlib==3.6.2
-numpy==1.24.0
-pandas==1.5.2
-pytest==7.2.0
-requests==2.31.0
-scipy==1.10.0
-setuptools==65.6.3
-shapely==2.0.0
-utm==0.7.0
-rasterio==1.3.6
-cchardet==2.1.7
-isoweek==1.3.3
-charset_normalizer==3.3.0
-mkdocs
-mkdocs-jupyter>=0.24.0
-mkdocs-material>=9.1.3 
-mkdocstrings
-mkdocstrings-crystal
