### 2023-10-18 030e72e
Add mkdocs-jupyter
