+++ .github/workflows/python-package.yml
@@ -28,6 +28,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
+        python -m pip install --upgrade Cython
         python -m pip install flake8 pytest
         pip install tox tox-gh-actions
     - name: Test with Tox
