+++ requirements.txt
@@ -12,6 +12,6 @@ setuptools==65.6.3
 shapely==2.0.0
 utm==0.7.0
 rasterio==1.3.6
-cchardet>=2.1.7
+faust-cchardet>=2.1.7
 isoweek==1.3.3
 charset_normalizer>=3.3.0
