### 2023-10-18 a3db866
Create CITATION.cff
