+++ README.md
@@ -219,7 +219,7 @@ Table generated by gtfs-segments using data from the San Francisco’s Muni syst
 
 Each row does not represent one segment. Rather, each row maps to a combination of a segment, a route that includes that segment, and a direction. For instance, a segment included in eight routes will appear as eight rows, which will have the same information except for route_id and traversals (since some routes might traverse the segment more than others). This choice enables filtering by route and preserves how many times each route traverses each segment during the measurement interval. The direction identifier is used for very rare cases (mostly loops) in which a route visits the same two stops, in the same order, but in different directions.
 ## Visualize Spacings
-
+> ℹ️ For more information on visualization refer to the [Visualization Tutorial](https://gtfs-segments.readthedocs.io/en/latest/tutorials/Visualization)
 ```python
 from gtfs_segments import view_spacings
 view_spacings(segments_df,route = '8',segment = '6364-3725-1',basemap=True)
@@ -241,7 +241,7 @@ Optionally save figure using
 ```python
 plot_hist(segments_df,file_path = "spacings_hist.png",save_fig = True)
 ```
-> ℹ️ For more information on visualization refer to the [Visualization Tutorial](https://gtfs-segments.readthedocs.io/en/latest/tutorials/Visualization)
+
 ## Get Summary Stats
 ```python
 from gtfs_segments import summary_stats
