### 2024-01-30 #6306 @xoolive
Apart from that minor comment, I can easily tick all the boxes. The objective of the library is simple but clear, and it does the job it promises. Good job!

Extra recommendation for code quality in the future, of course non-blocking for acceptance:

- [ ] use ruff for linting/formatting
- [ ] complete full annotations, check with mypy
