### 2024-01-31 667b85f
Provide defective trip count | Update Install in Readme
