### 2024-01-31 4822759
Fix partridge tests. Now uses only Ann Arbor
