+++ test/test_partridge.py
@@ -13,7 +13,7 @@ class TestPartridge(unittest.TestCase):
         self.gtfs_path = os.path.join(
             test_dir,
             "data",
-            "San Francisco-San Francisco Municipal Transportation Agency (SFMTA, Muni)-CA",
+            "Ann Arbor-University of Michigan Transit Services-MI",
             "gtfs.zip",
         )
 
@@ -27,45 +27,48 @@ class TestPartridge(unittest.TestCase):
         assert feed.stops.crs == "epsg:4326"
         assert ["shape_id", "geometry"] == list(feed.shapes.columns)
         assert [
-            "stop_code",
-            "stop_url",
             "stop_id",
-            "stop_desc",
+            "stop_code",
             "stop_name",
-            "location_type",
+            "stop_desc",
             "zone_id",
+            "stop_url",
+            "location_type",
+            "parent_station",
+            "stop_timezone",
+            "wheelchair_boarding",
             "geometry",
         ] == list(feed.stops.columns)
 
     def test_service_ids_by_date(self):
         service_ids_by_date = ptg.read_service_ids_by_date(self.gtfs_path)
 
-        assert len(service_ids_by_date) == 161
-        assert datetime.date(2022, 7, 30) in service_ids_by_date
+        assert len(service_ids_by_date) == 133
+        assert datetime.date(2022, 2, 17) in service_ids_by_date
 
     def test_dates_by_service_ids(self):
         dates_by_service_ids = ptg.read_dates_by_service_ids(self.gtfs_path)
-        assert len(dates_by_service_ids) == 6
-        assert frozenset({"2_merged_11239965"}) in dates_by_service_ids
+        assert len(dates_by_service_ids) == 16
+        assert frozenset({"2"}) in dates_by_service_ids
 
     def test_trip_counts_by_date(self):
         trip_counts_by_date = ptg.read_trip_counts_by_date(self.gtfs_path)
-        assert len(trip_counts_by_date) == 161
-        assert datetime.date(2022, 8, 20) in trip_counts_by_date
+        assert len(trip_counts_by_date) == 133
+        assert datetime.date(2022, 4, 12) in trip_counts_by_date
 
     def test_busiest_date(self):
         date, service_ids = ptg.read_busiest_date(self.gtfs_path)
-        assert date == datetime.date(2022, 10, 3)
-        assert service_ids == frozenset({"1_merged_11239966"})
+        assert date == datetime.date(2022, 1, 10)
+        assert service_ids == frozenset({"7"})
 
     def test_busiest_week(self):
         service_ids_by_date = ptg.read_busiest_week(self.gtfs_path)
         assert service_ids_by_date == {
-            datetime.date(2022, 10, 3): frozenset({"1_merged_11239966"}),
-            datetime.date(2022, 10, 4): frozenset({"1_merged_11239966"}),
-            datetime.date(2022, 10, 5): frozenset({"1_merged_11239966"}),
-            datetime.date(2022, 10, 6): frozenset({"1_merged_11239966"}),
-            datetime.date(2022, 10, 7): frozenset({"1_merged_11239966"}),
-            datetime.date(2022, 10, 8): frozenset({"2_merged_11239968"}),
-            datetime.date(2022, 10, 9): frozenset({"3_merged_11239967"}),
+            datetime.date(2022, 1, 10): frozenset({"7"}),
+            datetime.date(2022, 1, 11): frozenset({"10"}),
+            datetime.date(2022, 1, 12): frozenset({"10"}),
+            datetime.date(2022, 1, 13): frozenset({"10"}),
+            datetime.date(2022, 1, 14): frozenset({"4"}),
+            datetime.date(2022, 1, 15): frozenset({"11"}),
+            datetime.date(2022, 1, 16): frozenset({"9"}),
         }
