### 2024-01-31 37592fe
Linting with RUFF
