+++ gtfs_segments/partridge_mod/__init__.py
@@ -6,10 +6,11 @@ from .readers import (
     load_raw_feed,
     read_busiest_date,
     read_busiest_week,
-    read_service_ids_by_date,
     read_dates_by_service_ids,
+    read_service_ids_by_date,
     read_trip_counts_by_date,
 )
+
 # from .writers import extract_feed
 
 
