+++ pyproject.toml
@@ -26,7 +26,7 @@ maintainers = [
 ]
 readme = "README.MD"
 license = {file = "LICENSE", name = "MIT License"}
-keywords=["Python", "GTFS", "Transit", "Stop Spacings", "GIS"],
+keywords=["Python", "GTFS", "Transit", "Stop Spacings", "GIS"]
 classifiers = [
     "Development Status :: 5 - Stable",
     "License :: OSI Approved :: MIT License",
