+++ pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = 'GTFS Segments'
+name = 'gtfs_segments'
 # Support Python 3.9+.
 requires-python = ">=3.9"
 version = "2.0.3"
