+++ tox.ini
@@ -1,28 +1,0 @@
-[tox]
-env_list =
-    py{36,37,38,39,310}
-minversion = 4.11.3
-
-[gh-actions]
-python =
-    3.6: py36, flake8
-    3.7: py37
-    3.8: py38
-    3.9: py39
-    3.10: py310
-
-[testenv]
-description = run the tests with unittest
-deps = -rrequirements.txt
-commands =
-    python -m unittest
-
-[testenv:lint]
-description = run linters
-skip_install = true
-deps = flake8==6.1.0
-commands = 
-    # stop the build if there are Python syntax errors or undefined names
-    flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
-    # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-    flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
