### 2024-02-01 811b362
Added more type hints | Checks with mypy
