### 2024-02-01 077b1dc
Some more type hints
