+++ gtfs_segments/route_stats.py
@@ -1,16 +1,17 @@
 from datetime import timedelta
-from typing import Any, Callable, Dict, List
+from typing import Any, Dict, List
 
-import matplotlib.cm as cm
 import numpy as np
 import pandas as pd
 import utm
 from numpy.typing import NDArray
+from shapely.geometry import LineString
 
 from .geom_utils import code
+from .partridge_mod.gtfs import Feed
 
 
-def get_zone_epsg_from_ls(geom) -> int:
+def get_zone_epsg_from_ls(geom: LineString) -> int:
     """
     > The function takes a dataframe with a geometry column and returns the EPSG code for the UTM zone
     that contains the geometry
@@ -27,7 +28,7 @@ def get_zone_epsg_from_ls(geom) -> int:
     return code(zone, lat)
 
 
-def get_sec(time_str) -> int:
+def get_sec(time_str: str) -> int:
     """
     It takes a string in the format of hh:mm:ss and returns the number of seconds
 
@@ -41,7 +42,7 @@ def get_sec(time_str) -> int:
     return int(h) * 3600 + int(m) * 60 + int(s)
 
 
-def get_trips_len(df, time) -> int:
+def get_trips_len(df: pd.DataFrame, time: int) -> int:
     """
     > It returns the number of trips that are currently active at a given time
 
@@ -55,7 +56,7 @@ def get_trips_len(df, time) -> int:
     return len(df[(df.start_time <= time) & (df.end_time >= time)].trip_id.unique())
 
 
-def get_trips(df, time) -> pd.DataFrame:
+def get_trips(df: pd.DataFrame, time: int) -> pd.DataFrame:
     """
     > Return all trips that are currently active at a given time
 
@@ -69,7 +70,7 @@ def get_trips(df, time) -> pd.DataFrame:
     return df[(df.start_time <= time) & (df.end_time >= time)]
 
 
-def get_peak_time(df) -> List:
+def get_peak_time(df: pd.DataFrame) -> List:
     """
     The function `get_peak_time` takes a dataframe of bus trips and returns the number of buses and the
     time at which the most buses are running.
@@ -93,7 +94,7 @@ def get_peak_time(df) -> List:
     return [best, peak_time]
 
 
-def get_service_length(df) -> int:
+def get_service_length(df: pd.DataFrame) -> int:
     """
     The function takes a GTFS feed and a route_id, sorts the dataframe by stop_sequence, adjusts the
     shape_dist_traveled values if necessary, and returns the maximum shape_dist_traveled value.
@@ -112,7 +113,7 @@ def get_service_length(df) -> int:
     return df.shape_dist_traveled.max()
 
 
-def get_route_grp(route_df) -> pd.DataFrame:
+def get_route_grp(route_df: pd.DataFrame) -> pd.DataFrame:
     """
     It takes a dataframe of route information and returns a dataframe of route information with the
     first and last stop times for each trip
@@ -144,7 +145,7 @@ def get_route_grp(route_df) -> pd.DataFrame:
     return route_df_grp[col_subset]
 
 
-def get_all_peak_times(df_dir) -> Dict[str, NDArray[Any]]:
+def get_all_peak_times(df_dir: pd.DataFrame) -> Dict[str, NDArray[Any]]:
     """
     It takes a dataframe of bus trips and returns the peak number of buses and the time of the peak
 
@@ -182,32 +183,10 @@ def get_all_peak_times(df_dir) -> Dict[str, NDArray[Any]]:
     return {"peak_buses": peak_time_condensed}
 
 
-def get_cmap_string(palette, domain) -> Callable:
+def get_average_headway(df_dir: pd.DataFrame) -> Dict[str, float]:
     """
-    It takes a list of colors and a list of values, and returns a function that maps each value to its
[…]
