### 2024-02-06 9648431
Added function documentation
