### 2024-02-07 a54dc86
heatmap for any column, fetch feeds outside US
