+++ paper/paper.bib
@@ -61,7 +61,7 @@
 
 @software{Toso2023,
   title     = {gtfs_functions: Package with useful functions to create geo-spatial visualizations from a GTFS.},
-  author    = {Santiago Toso},
+  author    = {Toso, Santiago and Oja, Riku},
   month     = september,
   year      = 2023,
   publisher = {GitHub},
