+++ gtfs_segments/geom_utils.py
@@ -1,6 +1,5 @@
 from concurrent.futures import ThreadPoolExecutor
-from typing import Any, List, Tuple, Optional
-from click import style
+from typing import Any, List, Optional, Tuple
 
 import contextily as cx
 import folium
@@ -164,32 +163,43 @@ def view_spacings(
     elif level == "route":
         markersize = 40
         assert "route" in kwargs, "Please provide a route_id in route attibute"
-        gdf = gdf[gdf.route_id == kwargs["route"]].copy()
+        gdf = gdf[gdf.route_id.isin(kwargs["route"])].copy()
     elif level == "segment":
         markersize = 60
         assert "segment" in kwargs, "Please provide a segment_id in segment attibute"
-        gdf = gdf[gdf.segment_id == kwargs["segment"]].copy()
+        gdf = gdf[gdf.segment_id.isin(kwargs["segment"])].copy()
     else:
         raise ValueError("level must be either whole, route, or segment")
 
     # Plot the spacings
     if "route" in kwargs:
-        gdf = gdf[gdf.route_id == kwargs["route"]].copy()
-        ax = gdf.plot(
-            ax=ax,
-            linewidth=1.5,
-            color="#2ecc71",
-            label="Route ID:" + kwargs["route"],
-            zorder=2,
-        )
+        gdf = gdf[gdf.route_id.isin(kwargs["route"])].copy()
+        if len(kwargs["route"]) > 1:
+            ax = gdf.plot(
+                ax=ax,
+                linewidth=2,
+                column="route_id",
+                label="Route ID:" + str(kwargs["route"]),
+                zorder=2,
+                cmap="tab20",
+                legend=True,
+            )
+        else:
+            ax = gdf.plot(
+                ax=ax,
+                linewidth=2,
+                color="#2ecc71",
+                label="Route ID:" + str(kwargs["route"]),
+                zorder=2,
+            )
     if "segment" in kwargs:
         try:
-            gdf = gdf[gdf.segment_id == kwargs["segment"]].copy()
+            gdf = gdf[gdf.segment_id.isin(kwargs["segment"])].copy()
         except ValueError as e:
             raise ValueError(f"No such segment exists. Check if direction_id is incorrect {e}")
         ax = gdf.plot(
             ax=ax,
-            linewidth=2,
+            linewidth=2.5,
             color="#000000",
             label="Segment ID: " + str(kwargs["segment"]),
             zorder=3,
@@ -211,7 +221,10 @@ def view_spacings(
         df = gpd.GeoDataFrame(gdf, crs=crs)
         cx.add_basemap(ax, crs=df.crs, source=map_provider, attribution_size=5)
     plt.axis(axis)
-    plt.legend(loc="lower right")
+    if level != "segment":
+        plt.legend(loc="best")
+    else:
+        ax.legend().set_visible(False)
     return ax
 
 
@@ -624,7 +637,13 @@ def nearest_points_parallel(stop_df: gpd.GeoDataFrame, k_neighbors: int = 5) -> 
     return stop_df
 
 
-def view_heatmap(gdf: gpd.GeoDataFrame, cmap: Optional[str] = "RdYlBu", light_mode : bool = True, interactive:bool = False) -> Any:
+def view_heatmap(
+    gdf: gpd.GeoDataFrame,
+    column: str = "distance",
+    cmap: Optional[str] = "RdYlBu",
+    light_mode: bool = True,
+    interactive: bool = False,
+) -> Any:
     """
     Generates a heatmap visualization of a GeoDataFrame.
 
@@ -638,39 +657,71 @@ def view_heatmap(gdf: gpd.GeoDataFrame, cmap: Optional[str] = "RdYlBu", light_mo
         Any: The generated heatmap visualization.
 
     """
-    MAX_RANGE = gdf["distance"].max()
[…]
