+++ paper/paper.bib
@@ -1,6 +1,6 @@
 @misc{devunuri2023bus,
   title         = {Bus Stop Spacings Statistics: Theory and Evidence},
-  author        = {Saipraneeth Devunuri and Shirin Qiam and Lewis Lehe and Ayush Pandey and Dana Monzer},
+  author        = {Devunuri, Saipraneeth  and Qiam, Shirin  and Lehe, Lewis  and Pandey, Ayush  and Monzer, Dana },
   year          = {2023},
   eprint        = {2208.04394},
   archiveprefix = {arXiv},
@@ -10,7 +10,7 @@
 
 @misc{devunuri2023chatgpt,
   title         = {ChatGPT for GTFS: From Words to Information},
-  author        = {Saipraneeth Devunuri and Shirin Qiam and Lewis Lehe},
+  author        = {Devunuri, Saipraneeth  and Qiam, Shirin  and Lehe, Lewis},
   year          = {2023},
   eprint        = {2308.02618},
   archiveprefix = {arXiv},
@@ -19,7 +19,7 @@
 }
 
 @data{DVN/SFBIVU_2022,
-  author    = {Devunuri, Saipraneeth and Shirin Qiam and Lewis Lehe},
+  author    = {Devunuri, Saipraneeth and Qiam, Shirin  and Lehe, Lewis },
   publisher = {Harvard Dataverse},
   title     = {{Bus Stop Spacings for Transit Providers in the US}},
   unf       = {UNF:6:43Qgi8ldmVm8XIxuC0EoKA==},
@@ -70,7 +70,7 @@
 
 @software{Whalen2023,
   title     = {partridge: A fast, forgiving GTFS reader built on pandas DataFrames},
-  author    = {Danny Whalen},
+  author    = {Whalen, Danny},
   month     = september,
   year      = 2023,
   publisher = {GitHub},
@@ -115,4 +115,15 @@
   journal = {arXiv preprint cs/9901013},
   year    = {1999},
   doi     = {10.48550/arXiv.cs/9901013}
+}
+
+@inproceedings{levenshtein1965binary,
+  title={Binary codes capable of correcting deletions, insertions, and reversals},
+  author={Levenshtein, Vladimir Iosifovich},
+  booktitle={Doklady Akademii Nauk},
+  volume={163},
+  number={4},
+  pages={845--848},
+  year={1965},
+  organization={Russian Academy of Sciences}
 }
