+++ requirements.txt
@@ -15,3 +15,4 @@ rasterio==1.3.6
 faust-cchardet>=2.1.7
 isoweek==1.3.3
 charset_normalizer>=3.3.0
+thefuzz>=0.22.1
