+++ _config.yml
@@ -1,1 +1,2 @@
 theme: jekyll-theme-cayman
+exclude: [paper]
