### 2024-02-08 f2e2d85
typo fixes, paper edits
