+++ paper/paper.md
@@ -44,20 +44,25 @@ The `gtfs-segments` package has four main functionalities: (1) Acquiring GTFS fe
 
 ## Acquiring GTFS feeds
 
-GTFS feeds undergo frequent changes, and obtaining the latest feed is essential to calculate up-to-date stop spacings. The package provides a convenient functionality to source and download over 1100 GTFS active feeds from across the world (over 550 in the US and 80 in Canada) from Mobility Database Catalogs[@MobData2023]. The package includes keyword search functionality to search for GTFS feeds using the name of the provider (full name or abbreviation) or the place (city or state (applicable only for the US)). Additionally, to account for typographical errors, the package also implements a fuzzy search feature by matching sub-strings based on [Levenshtein distance](https://en.wikipedia.org/wiki/Levenshtein_distance) algorithm, which is used popularly in text auto-correct.
+GTFS feeds undergo frequent changes, and obtaining the latest feed is essential to calculate up-to-date stop spacings. The package provides a convenient functionality to source and download over 1100 active GTFS feeds from across the world (over 550 in the US and 80 in Canada) from Mobility Database Catalogs[@MobData2023]. The package includes keyword search functionality to search for GTFS feeds using either the name of the provider (full name or abbreviation) or the place (city or state[^1]). Additionally, to account for typographical errors, the package also implements a fuzzy search feature by matching sub-strings based on Levenshtein distance[@levenshtein1965binary] algorithm, which is used popularly in text auto-correct.
+
+[^1]: Search by state is currently applicable only to the US alone.
 
 ## Computing segments
 
 A `segment` is defined by three elements: (i) a start stop, (ii) an end stop and (iii) the path that the bus travels along the route in between the two consecutive stops. The segments are computed using route shape geometries and stop locations included in GTFS data. Packages such as `gtfs2gps`[@pereira2023exploring] and `gtfs_functions`[@Toso2023] can compute segments, but this package does so in computationally-efficient way and includes a new way to account for *errors* in the GTFS feed.
 
[…]
