+++ paper/paper.bib
@@ -56,7 +56,8 @@
   title   = {Bus stop spacing with heterogeneous trip lengths and elastic demand},
   author  = {Lehe, Lewis and Pandey, Ayush},
   journal = {Available at SSRN 4135394},
-  doi     = {10.2139/ssrn.4135394}
+  doi     = {10.2139/ssrn.4135394},
+  year    = {2022},
 }
 
 @software{Toso2023,
@@ -125,5 +126,6 @@
   number={4},
   pages={845--848},
   year={1965},
-  organization={Russian Academy of Sciences}
+  organization={Russian Academy of Sciences},
+  url= {https://nymity.ch/sybilhunting/pdf/Levenshtein1966a.pdf}
 }
