+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'GTFS Segments: A fast and efficient library to generate bus stop spacings'
+title: 'GTFS Segments: A Fast and Efficient Library to Generate Bus Stop Spacings'
 tags:
   - Python
   - GTFS
@@ -25,18 +25,18 @@ bibliography: paper.bib
 header-includes: |
     \usepackage{caption}
     \usepackage{subcaption}
+    \usepackage{graphicx}
 ---
 
 # Summary
 
-The GTFS Segments (gtfs-segments) library is an open-source Python toolkit for computing, visualizing and analyzing bus stop spacings: the distance a transit bus travels between stops. The library reads General Transit Feed Specification (GTFS) data, snaps bus stops to points along routes, divides routes into segments (the pieces of routes between two stops), and then produces a `GeoDataFrame` containing information about each segment. The library also features several functions that act on this GeoDataFrame. It can produce summary statistics of the spacings for a given network, using various weighting schemes (i.e., weighting by frequency of service), as well as histograms of spacings that display their full distribution. In addition to network-level statistics, the package can also compute statistics for each route, such as its length, headway, speed, and average stop spacing. It can draw maps of networks, routes, or segments over a basemap---which allows the user to manually validate data. The segments DataFrame can be exported to `.csv` and `.geojson` files. The package can fetch the most up-to-date GTFS data from Mobility Data[@MobData2023] repositories for user convenience.
-
+The GTFS Segments (gtfs-segments) library is an open-source Python toolkit for computing, visualizing and analyzing bus stop spacings: the distance a transit bus travels between stops. The library reads General Transit Feed Specification (GTFS) data, snaps bus stops to points along routes, divides routes into segments (the pieces of routes between two stops), and then produces a `GeoDataFrame` containing information about each segment. The library also features several functions that act on this GeoDataFrame. It can produce summary statistics of the spacings for a given network, using various weighting schemes (i.e., weighting by frequency of service), as well as histograms of spacings that display their full distribution. In addition to network-level statistics, the package can also compute statistics for each route, such as its length, headway, speed, and average stop spacing. It can draw maps of networks, routes, or segments over a basemap---which allows the user to manually validate data. The segments DataFrame can be exported to `.csv` and `.geojson` files. The package can fetch the most up-to-date GTFS data from Mobility Data [@MobData2023] repositories for user convenience.
 
 # Statement of need
 
[…]
