+++ paper/paper.bib
@@ -1,6 +1,6 @@
 @misc{devunuri2023bus,
-  title         = {Bus Stop Spacings Statistics: Theory and Evidence},
-  author        = {Devunuri, Saipraneeth  and Qiam, Shirin  and Lehe, Lewis  and Pandey, Ayush  and Monzer, Dana },
+  title         ={ {Bus Stop Spacings Statistics: Theory and Evidence}},
+  author        = {Devunuri, Saipraneeth and Qiam, Shirin  and Lehe, Lewis  and Pandey, Ayush  and Monzer, Dana },
   year          = {2023},
   eprint        = {2208.04394},
   archiveprefix = {arXiv},
@@ -9,8 +9,8 @@
 }
 
 @misc{devunuri2023chatgpt,
-  title         = {ChatGPT for GTFS: From Words to Information},
-  author        = {Devunuri, Saipraneeth  and Qiam, Shirin  and Lehe, Lewis},
+  title         = {{ChatGPT for GTFS: From Words to Information}},
+  author        = {Devunuri, Saipraneeth and Qiam, Shirin  and Lehe, Lewis},
   year          = {2023},
   eprint        = {2308.02618},
   archiveprefix = {arXiv},
