+++ paper/paper.md
@@ -30,7 +30,7 @@ header-includes: |
 
 # Summary
 
-The GTFS Segments (gtfs-segments) library is an open-source Python toolkit for computing, visualizing and analyzing bus stop spacings: the distance a transit bus travels between stops. The library reads General Transit Feed Specification (GTFS) data, snaps bus stops to points along routes, divides routes into segments (the pieces of routes between two stops), and then produces a `GeoDataFrame` containing information about each segment. The library also features several functions that act on this GeoDataFrame. It can produce summary statistics of the spacings for a given network, using various weighting schemes (i.e., weighting by frequency of service), as well as histograms of spacings that display their full distribution. In addition to network-level statistics, the package can also compute statistics for each route, such as its length, headway, speed, and average stop spacing. It can draw maps of networks, routes, or segments over a basemap---which allows the user to manually validate data. The segments DataFrame can be exported to `.csv` and `.geojson` files. The package can fetch the most up-to-date GTFS data from Mobility Data [@MobData2023] repositories for user convenience.
+The GTFS Segments (gtfs-segments) library is an open-source Python toolkit for computing, visualizing and analyzing bus stop spacings: the distance a transit bus travels between stops. The library reads General Transit Feed Specification (GTFS) data, snaps bus stops to points along routes, divides routes into segments (the pieces of routes between two stops), and then produces a `GeoDataFrame` containing information about each segment. The library also features several functions that act on this GeoDataFrame. It can produce summary statistics of the spacings for a given network, using various weighting schemes (i.e., weighting by frequency of service), as well as histograms of spacings that display their full distribution. In addition to network-level statistics, the package can also compute statistics for each route, such as its length, headway, speed, and average stop spacing. It can draw maps of networks, routes, or segments over a basemap---which allows the user to manually validate data. The segments DataFrame can be exported to `.csv` and `.geojson` file formats. The package can fetch the most up-to-date GTFS data from Mobility Data [@MobData2023] repositories for user convenience.
 
 # Statement of need
 
@@ -40,23 +40,40 @@ The choice of bus stop spacing involves a tradeoff between accessibility and spe
 
 # Functionality
 
-The `gtfs-segments` package has four main functionalities: (1) Acquiring GTFS feeds (2) Computing segments (3) Visualizing stop spacings (4) Calculating stop spacing summary statistics. Each of these functionalities is further detailed below.
+The `gtfs-segments` package has four main functionalities: (1) Downloading GTFS feeds (2) Computing segments (3) Visualizing stop spacings (4) Calculating stop spacing summary statistics. Each is further detailed below.
 
-## Acquiring GTFS feeds
+## Downloading GTFS feeds
 
-GTFS feeds undergo frequent changes, and obtaining the latest feed is essential to calculate up-to-date stop spacings. The package facilitates accessing over 1100 active GTFS feeds from across the world from Mobility Database Catalogs [@MobData2023]. It includes keyword search functionality to search for GTFS feeds using either the name of the provider (full name or abbreviation) or the place (city or state[^1]). Additionally, to account for typographical errors, the package also implements a fuzzy search feature by matching sub-strings based on the Levenshtein distance [@levenshtein1965binary] algorithm.
-
-[^1]: Search by state is currently applicable only to the US alone.
+The package permits the user to search and download recent GTFS feeds from the Mobility Database Catalogs [@MobData2023]. It allows for keyword and fuzzy search of GTFS feeds using location (e.g., `Minneapolis`, `San Francisco`) or provider name (e.g., `WMATA`, `Capital Metro`) as input.
 
[…]
