+++ paper/paper.md
@@ -40,14 +40,14 @@ The choice of bus stop spacing involves a tradeoff between accessibility and spe
 
 # Functionality
 
-The `gtfs-segments` package has four main functionalities: (1) Downloading GTFS feeds (2) Computing segments (3) Visualizing stop spacings (4) Calculating stop spacing summary statistics. Each is further detailed below.
+`gtfs-segments` has four main functionalities: (1) Downloading GTFS feeds (2) Computing segments (3) Visualizing stop spacings (4) Calculating stop spacing summary statistics. Each is further detailed below.
 
 ## Downloading GTFS feeds
 
 The package permits the user to search and download recent GTFS feeds from the Mobility Database Catalogs [@MobData2023]. It allows for keyword and fuzzy search of GTFS feeds using location (e.g., `Minneapolis`, `San Francisco`) or provider name (e.g., `WMATA`, `Capital Metro`) as input.
 
 ## Computing segments
-The fundamental unit of analysis used by `gtfs-segments` is the *segment*, which is a piece of a bus network defined by three properties: (i) a start stop, (ii) an end stop and (iii) the path that the bus travels along the route in between the two consecutive stops. A segment's *spacing* is the distance of (iii). `gtfs-segments` splits a bus network into segments efficiently. This is a challenge because GTFS does not give the locations of stops as points along routes. Stops have coordinates *near* routes, but not exactly on them. \autoref{fig:snapping_difficulty} provides example cases where a stop is equidistant from multiple points on a route. Here, projecting the stop onto the route or snapping to the nearest geo-coordinate (lat, lon) may produce errors, such as stops that are out-of-order or stops being snapped far from their locations. Also, the time complexity of projection or snapping using brute force is $O(nm)$ for `n` stops and `m` geo-coordinates that represent the route shape.
+The fundamental unit of analysis used by `gtfs-segments` is the *segment*, which is a piece of a bus network defined by three properties: (i) a start stop, (ii) an end stop and (iii) the path that the bus travels along the route in between the two consecutive stops. A segment's *spacing* is the distance of (iii). `gtfs-segments` produces segments by efficiently and robustly snapping stop locations onto route shapes. \autoref{fig:snapping_difficulty} shows examples where a stop is equidistant from multiple route coordinates. Here, projecting the stop onto the route or snapping to the nearest geo-coordinate (lat, lon) may yield stops that are out-of-order or snapped far from their locations. Also, the time complexity of projection or snapping using brute force is $O(nm)$ for `n` stops and `m` geo-coordinates that represent the route shape.
 
 \begin{figure}[!h]
   \centering
@@ -56,7 +56,7 @@ The fundamental unit of analysis used by `gtfs-segments` is the *segment*, which
   \label{fig:snapping_difficulty}
 \end{figure}
 
[…]
