### 2024-02-09 #6306 @praneethd7
> Hi, just starting to have a look at your repo, I am seeing
>
> ```
> conda config --env --set channel_priority strict
> ```
>
> Honestly, this is usually fine, but it has been a mess every time I recommended that to Windows users. I would recommend to edit the block to
>
> ```
> conda create -n geo_env -c conda-forge python=3.11 geopandas
> ```
>
> Other settings are intrusive anyway.
>
> Also you may want to consider to submit the repo to conda-forge, it is rather easy and would cut the explanations short...

Thank you @xoolive for your time and review. I have changed the install instructions accordingly and verified they work.
