### 2024-02-09 #6306 @praneethd7
> Apart from that minor comment, I can easily tick all the boxes. The objective of the library is simple but clear, and it does the job it promises. Good job!
>
> Extra recommendation for code quality in the future, of course non-blocking for acceptance:
>
> * [ ]  use ruff for linting/formatting
> * [ ]  complete full annotations, check with mypy

Thank you for suggesting these code changes. I was initially using Black for formatting but Ruff is convenient and much better. I have formatted the code with Ruff. Also, I added type annotations for the whole repository and checked it with mypy. Thanks for introducing me to mypy. It is game changer
* [x]  use ruff for linting/formatting
* [x]  complete full annotations, check with mypy
