+++ paper/paper.md
@@ -22,10 +22,6 @@ affiliations:
 
 date: 25 September 2023
 bibliography: paper.bib
-header-includes: |
-    \usepackage{caption}
-    \usepackage{subcaption}
-    \usepackage{graphicx}
 ---
 
 # Summary
@@ -47,102 +43,30 @@ The choice of bus stop spacing involves a tradeoff between accessibility and spe
 The package permits the user to search and download recent GTFS feeds from the Mobility Database Catalogs [@MobData2023]. It allows for keyword and fuzzy search of GTFS feeds using location (e.g., `Minneapolis`, `San Francisco`) or provider name (e.g., `WMATA`, `Capital Metro`) as input.
 
 ## Computing segments
+
 The fundamental unit of analysis used by `gtfs-segments` is the *segment*, which is a piece of a bus network defined by three properties: (i) a start stop, (ii) an end stop and (iii) the path that the bus travels along the route in between the two consecutive stops. A segment's *spacing* is the distance of (iii). `gtfs-segments` produces segments by efficiently and robustly snapping stop locations onto route shapes. \autoref{fig:snapping_difficulty} shows examples where a stop is equidistant from multiple route coordinates. Here, projecting the stop onto the route or snapping to the nearest geo-coordinate (lat, lon) may yield stops that are out-of-order or snapped far from their locations. Also, the time complexity of projection or snapping using brute force is $O(nm)$ for `n` stops and `m` geo-coordinates that represent the route shape.
 
-\begin{figure}[!h]
-  \centering
-  \includegraphics[width=\textwidth]{snapping_difficulty.jpg}
-  \caption{Example route shapes with stop locations that are equidistant from multiple points along the route.}
-  \label{fig:snapping_difficulty}
-\end{figure}
+![Example route shapes with stop locations that are equidistant from multiple points along the route.\label{fig:snapping_difficulty}](snapping_difficulty.jpg){ width=80% }
 
-`gtfs-segments` overcomes these challenges by increasing the route resolution (i.e., adding points in between geo-coordinates), using spatial k-d trees, and using more than one nearest neighbor. The increase in resolution allows stops to be snapped to nearby points. Using k-d trees reduces the time complexity to $O(nlog(m))$ and makes it possible to compare among several snapping points without added computation. \autoref{fig:interpolate} shows an example where initially snapping to the nearest point produces out-of-order stops (3/4/2) and stop 5 is snapped far away from its location. Increasing the resolution (second panel) fixes 5's location problem but the ordering problem persists. By using `k=3` nearest neighbors, we find a proper ordering (last panel). Once every stop has been snapped to a geo-coordinate on the route shape, the shape is segmented between stops and each segment's geometry is stored in a GeoDataFrame. 
[…]
