+++ paper/heatmap_and_histogram.tex
@@ -1,0 +1,20 @@
+\documentclass[preview]{standalone}
+\usepackage{subcaption}
+\usepackage{graphicx}
+\begin{document}
+\begin{figure}[!h]
+    \centering
+    \begin{subfigure}[]{0.52\textwidth}
+      \centering
+      \includegraphics[width=\textwidth]{heatmap_interactive.jpg}
+      \caption{Interactive Heatmap}
+      \label{fig:div}
+      \end{subfigure}
+    \begin{subfigure}[]{0.42\textwidth}
+      \centering
+      \includegraphics[width=\textwidth]{hist.jpg}
+      \caption{Histogram of stop spacings}
+      \label{fig:hist}
+    \end{subfigure}
+\end{figure}
+\end{document}
