### 2024-02-09 #6306 @praneethd7
> Overall, I found the work clear and complete. I have a minor feedback regarding the state of the field: authors clearly stated why this software is needed in practice though it is not clear if there are any previous contributions/packages similar to their study and what is their novelty when compared to those. There are some packages mentioned in the acknowledgement field. Authors are referring that their study inspired from those. The level of inspiration and the improvement of this study over those studies can be an input to tackle the lack of "state of the field" section.

Thank you @nagellette for your review and time.  To provide more information about the state of the field, we made the following changes:
- We added a new section termed "Functionality" and explained the different functionalities such as downloading feeds, visualization, and summary statistics the package provides. Consequently, we made the "computing segments" section into a subsection under Functionality
- Besides computing segments, other features currently do not exist in other packages. We added the following statement in the manuscript:

> Packages such as `gtfs2gps` and `gtfs_functions` also compute segments. In addition to its snapping algorithm, visualization, download, and statistical functionalities, `gtfs-segments` is distinguished from those in two ways. First, it has a faster processing rate[^1] to compute segments both with and without parallel processing (see Table 1). Second, `gtfs-segments` is tolerant to deviations from GTFS standards. For example, because the Chicago Transit Authority does not have an agency_id in its routes.txt, `gtfs2gps` fails to read it even though this field is not needed for obtaining segments.

- To provide evidence that our package is indeed faster than those mentioned, we added Table 1 which provides a comparison of processing rates.
- We added two figures generated from the package: a heatmap and a histogram, in the visualization subsection.
