+++ .github/workflows/python-publish.yml
@@ -35,5 +35,5 @@ jobs:
     - name: Publish package
       uses: pypa/gh-action-pypi-publish@27b31702a0e7fc50959f5ad993c78deac1bdfc29
       with:
-        user: ${{secrets.TWINE_USERNAME}}
-        password: ${{ secrets.TWINE_PASSWORD }}
+        user: __token__
+        password: ${{ secrets.TEST_PYPI_API_TOKEN }}
