+++ .github/workflows/draft-pdf.yml
@@ -1,4 +1,8 @@
-on: [push]
+on:
+  push:
+    branches: [ "main" ]
+    paths:
+      - 'paper/**'
 
 jobs:
     paper:
