+++ .github/workflows/python-publish.yml
@@ -11,7 +11,11 @@ name: Upload Python Package
 on:
   release:
     types: [published]
-
+  push:
+    branches: [ "main" ]
+    paths:
+      - '.github/workflows/python-publish.yml'
+      
 permissions:
   contents: read
 
