+++ pyproject.toml
@@ -1,5 +1,6 @@
 [project]
 name = 'gtfs_segments'
+description = "GTFS Segments"
 # Support Python 3.9+.
 requires-python = ">=3.9"
 version = "2.1.0"
@@ -26,7 +27,7 @@ authors = [
 maintainers = [
     {name = "Saipraneeth Devunuri", email = "sd37@illinois.edu"}
 ]
-readme = "README.MD"
+readme = {file = "README.md", content-type = "text/markdown"}
 license = {file = "LICENSE", name = "MIT License"}
 keywords=["Python", "GTFS", "Transit", "Stop Spacings", "GIS"]
 classifiers = [
