### 2024-02-19 #6306 @oliviaguest
@praneethd7 do you think maybe the figure text can be a little more legible? otherwise the paper looks good to me
