### 2024-02-19 dd6288e
Enlarge figures JOSS
