+++ paper/heatmap_and_histogram.tex
@@ -4,13 +4,13 @@
 \begin{document}
 \begin{figure}[!h]
     \centering
-    \begin{subfigure}[]{0.52\textwidth}
+    \begin{subfigure}[]{0.75\textwidth}
       \centering
       \includegraphics[width=\textwidth]{heatmap_interactive.jpg}
       \caption{Interactive Heatmap}
       \label{fig:div}
       \end{subfigure}
-    \begin{subfigure}[]{0.42\textwidth}
+    \begin{subfigure}[]{0.8\textwidth}
       \centering
       \includegraphics[width=\textwidth]{hist.jpg}
       \caption{Histogram of stop spacings}
