### 2024-02-19 c090b50
enlarge all figures
