### 2024-02-19 #6306 @praneethd7
Thank you @oliviaguest for the suggestion. I have enlarged Figures 2 and 3

### Additional Author Tasks After Review is Complete

> Latest version is [v2.1.4](https://github.com/UTEL-UIUC/gtfs_segments/releases/tag/v2.1.4)

> DOI : [10.5281/zenodo.10681151](https://doi.org/10.5281/zenodo.10681151)
