### 2024-02-27 b6560b1
Speed division - np.inf
