+++ README.md
@@ -222,8 +222,10 @@ Table generated by gtfs-segments using data from San Francisco’s Muni system. 
 4. `route_id`: The same route ID listed in the agency's routes.txt file.
 5. `direction_id`: The route's direction identifier.
 6. `traversals`: The number of times the indicated route traverses the segment during the "measurement interval." The "measurement interval" chosen is the busiest day in the GTFS schedule: the day which has the most bus services running.
-7. `distance`: The length of the segment in meters.
+7. `distance`: The length of the bus segment in meters.
 8. `geometry`: The segment's LINESTRING (a format for encoding geographic paths). All geometries are re-projected onto Mercator (EPSG:4326/WGS84) to maintain consistency.
+9. `traversal_time`: The time (in seconds) that it takes for the bus to traverse the segment.
+10. `speed`: The speed of the bus (in kmph) while traversing the segment. Default to `np.inf`♾ in case `traversal_time` is zero.
 
 Each row does not represent one segment. Rather, each row maps to a combination of a segment, a route that includes that segment, and a direction. For instance, a segment included in eight routes will appear as eight rows, which will have the same information except for route_id and traversals (since some routes might traverse the segment more than others). This choice enables filtering by route and preserves how many times each route traverses each segment during the measurement interval. The direction identifier is used for very rare cases (mostly loops) in which a route visits the same two stops, in the same order, but in different directions.
 
