### 2024-02-27 643f762
Exlcuding trips with no shape_id
