+++ pyproject.toml
@@ -3,7 +3,7 @@ name = 'gtfs_segments'
 description = "GTFS Segments: A fast and efficient library to generate bus stop spacings"
 # Support Python 3.9+.
 requires-python = ">=3.9"
-version = "2.1.4"
+version = "2.1.5"
 dependencies = [
     "geopandas",
     "scipy",
