### 2024-03-07 8e92735
update citation JOSS | min_spacing
