### 2024-03-11 #6306 @praneethd7
> Thank you @oliviaguest for the suggestion. I have enlarged Figures 2 and 3
>
> ### Additional Author Tasks After Review is Complete
> * [x]  Double check authors and affiliations (including ORCIDs)
> * [x]  Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
>
> > Latest version is [v2.1.4](https://github.com/UTEL-UIUC/gtfs_segments/releases/tag/v2.1.4)
>
> * [x]  Archive the release on Zenodo/figshare/etc and post the DOI here.
>
> > DOI : [10.5281/zenodo.10681151](https://doi.org/10.5281/zenodo.10681151)
>
> * [x]  Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
> * [x]  Make sure that the license listed for the archive is the same as the software license.

@oliviaguest Just thought of checking in as it has been about 3 weeks. Thank you!
