+++ gtfs_segments/__init__.py
@@ -1,7 +1,7 @@
 """
 The gtfs_segments package main init file.
 """
-__version__ = "2.1.5"
+import importlib.metadata
 from .geom_utils import view_heatmap, view_spacings, view_spacings_interactive
 from .gtfs_segments import get_gtfs_segments, pipeline_gtfs, process_feed
 from .mobility import (
@@ -13,6 +13,7 @@ from .partridge_func import get_bus_feed
 from .route_stats import get_route_stats
 from .utils import export_segments, plot_hist, process, summary_stats
 
+__version__ = importlib.metadata.version("gtfs_segments")
 __all__ = [
     "__version__",
     "get_gtfs_segments",
