+++ gtfs_segments/partridge_mod/parsers.py
@@ -18,7 +18,7 @@ DATE_FORMAT = "%Y%m%d"
 
 # Why 2^18? See https://git.io/vxB2P.
 @lru_cache(maxsize=2**18)
-def parse_time(val: str) -> int:
+def parse_time(val: str) -> np.float32:
     """
     The function `parse_time` takes a string representing a time value in the format "hh:mm:ss" and
     returns the equivalent time in seconds as a numpy int, or returns the input value if it is
@@ -28,19 +28,16 @@ def parse_time(val: str) -> int:
       val (str): The parameter `val` is a string representing a time value in the format "hh:mm:ss".
 
     Returns:
-      a value of type np.float64.
+      a value of type np.float32.
     """
-    if val is np.nan or isinstance(val, int):
+    if isinstance(val, float) or (isinstance(val, float) and np.isnan(val)):  # Corrected handling for np.nan
         return val
-    val = val.strip()
-    h, m, s = val.split(":")
-    ssm = int(h) * 3600 + int(m) * 60 + int(s)
-    if val == "":
+    if str(val) == "":
         return np.nan
+    val = str(val).strip()
 
-    # pandas doesn't have a NaN int, use floats
-    # return np.float64(ssm)
-    return ssm
+    h, m, s = map(int, val.split(":"))
+    return np.float32(h * 3600 + m * 60 + s)
 
 
 @lru_cache(maxsize=2**18)
@@ -61,14 +58,22 @@ def parse_date(val: str) -> datetime.date:
 
 
 @lru_cache(maxsize=2**18)
-def parse_numeric(val: Any) -> float:
+def parse_float(val: Any) -> float:
     try:
         return float(val)
     except ValueError:
         return np.nan
+    
+@lru_cache(maxsize=2**18)
+def parse_integer(val: Any) -> float:
+    try:
+        return int(val)
+    except ValueError:
+        return np.nan
 
 
-vparse_numeric = np.vectorize(parse_numeric)
+vparse_float = np.vectorize(parse_float)
+vparse_int = np.vectorize(parse_integer)
 vparse_time = np.vectorize(parse_time)
 vparse_date = np.vectorize(parse_date)
 
@@ -164,13 +169,13 @@ def transforms_dict() -> Dict[str, Dict[str, Any]]:
             "converters": {
                 "start_date": vparse_date,
                 "end_date": vparse_date,
-                "monday": vparse_numeric,
-                "tuesday": vparse_numeric,
-                "wednesday": vparse_numeric,
-                "thursday": vparse_numeric,
-                "friday": vparse_numeric,
-                "saturday": vparse_numeric,
-                "sunday": vparse_numeric,
+                "monday": vparse_float,
+                "tuesday": vparse_float,
+                "wednesday": vparse_float,
+                "thursday": vparse_float,
+                "friday": vparse_float,
+                "saturday": vparse_float,
+                "sunday": vparse_float,
             },
             "required_columns": (
                 "service_id",
@@ -189,7 +194,7 @@ def transforms_dict() -> Dict[str, Dict[str, Any]]:
             "usecols": {"service_id": "str", "date": "str", "exception_type": "int8"},
             "converters": {
                 "date": vparse_date,
-                "exception_type": vparse_numeric,
+                "exception_type": vparse_float,
             },
             "required_columns": ("service_id", "date", "exception_type"),
         },
@@ -203,9 +208,9 @@ def transforms_dict() -> Dict[str, Dict[str, Any]]:
                 "transfer_duration": "float16",
             },
             "converters": {
-                "price": vparse_numeric,
-                "payment_method": vparse_numeric,
-                "transfer_duration": vparse_numeric,
+                "price": vparse_float,
+                "payment_method": vparse_float,
[…]
