### 2024-03-17 f180347
include partridge_mod
