+++ pyproject.toml
@@ -1,9 +1,13 @@
+[build-system]
+requires = ["setuptools", "setuptools-scm"]
+build-backend = "setuptools.build_meta"
+
 [project]
 name = 'gtfs_segments'
 description = "GTFS Segments: A fast and efficient library to generate bus stop spacings"
 # Support Python 3.9+.
 requires-python = ">=3.9"
-version = "2.1.6"
+version = "2.1.7"
 dependencies = [
     "geopandas",
     "scipy",
@@ -41,7 +45,7 @@ classifiers = [
 ]
 
 [tool.setuptools]
-packages = ["gtfs_segments"]
+packages = ["gtfs_segments","gtfs_segments.partridge_mod"]
 
 [project.urls]
 Homepage = "https://github.com/UTEL-UIUC/gtfs_segments"
