+++ pyproject.toml
@@ -9,11 +9,11 @@ description = "GTFS Segments: A fast and efficient library to generate bus stop 
 requires-python = ">=3.9"
 version = "2.1.7"
 dependencies = [
-    "geopandas",
+    "geopandas >= 0.12.0",
     "scipy",
     "shapely",
-    "numpy",
-    "pandas",
+    "numpy >= 1.25.0",
+    "pandas >= 2.0.0",
     "matplotlib",
     "utm",
     "contextily",
