### 2024-03-19 #6306 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
